Do not allocate the buffer in decode_config.
decode_config would allocate a vector with a capacity of

input.len() * 4 / 3

and pass that to decode_config_buf which would resize the vector to

(input.len() + 7) / 8 * 6

Since decode_config_buf always resizes the vector to a more accurate
size the allocation performed in decode_config is unnecessary. For large
inputs it will allocate more space than necessary and for small inputs
(<5 bytes) it will cause two allocations.